refactor: extract standings domain
