+++ src/frontend/components/OverlayContainer/SessionBestLapUpdater.spec.tsx
@@ -5,13 +5,13 @@ import {
   useStandingsSettings,
   useRelativeSettings,
   useInformationBarSettings,
-} from '../Standings/hooks';
+} from '@irdashies/domain';
 import { SessionBestLapUpdater } from './SessionBestLapUpdater';
 
 vi.mock('@irdashies/context', () => ({
   SessionBestLapStoreUpdater: vi.fn(() => null),
 }));
-vi.mock('../Standings/hooks', () => ({
+vi.mock('@irdashies/domain', () => ({
   useStandingsSettings: vi.fn(),
   useRelativeSettings: vi.fn(),
   useInformationBarSettings: vi.fn(),
