+++ src/frontend/components/TrackMap/FlatTrackMap.spec.tsx
@@ -5,7 +5,7 @@ import { render } from '@testing-library/react';
 
 vi.mock('./hooks/useTrackId');
 vi.mock('./hooks/useDriverProgress');
-vi.mock('../Standings/hooks/useDriverLivePositions');
+vi.mock('@irdashies/domain/standings/useDriverLivePositions');
 vi.mock('./hooks/useFlatTrackMapSettings');
 vi.mock('./hooks/useHighlightColor');
 vi.mock('@irdashies/context', () => {
@@ -27,7 +27,7 @@ vi.mock('./tracks/tracks.json', () => ({
 
 import { useTrackId } from './hooks/useTrackId';
 import { useDriverProgress } from './hooks/useDriverProgress';
-import { useDriverLivePositions } from '../Standings/hooks/useDriverLivePositions';
+import { useDriverLivePositions } from '@irdashies/domain';
 import { useFlatTrackMapSettings } from './hooks/useFlatTrackMapSettings';
 import { useHighlightColor } from './hooks/useHighlightColor';
 import { useSessionVisibility, useTelemetryValue } from '@irdashies/context';
