+++ src/frontend/domain/standings/createStandings.spec.ts
