+++ src/frontend/domain/standings/createStandings.ts
