+++ src/frontend/domain/standings/useHighlightColor.tsx
@@ -4,4 +4,3 @@ export const useHighlightColor = () => {
   const { currentDashboard } = useDashboard();
   return currentDashboard?.generalSettings?.highlightColor ?? 960745;
 };
-
