+++ src/frontend/domain/standings/useInformationBarSettings.ts
