+++ src/frontend/domain/standings/useRelativeSettings.ts
