+++ src/frontend/utils/referenceLap.ts
@@ -1,0 +1,5 @@
+export const getBucketIndex = (
+  trackPct: number,
+  pointsCount: number
+): number =>
+  Math.min(Math.max(Math.floor(trackPct * pointsCount), 0), pointsCount - 1);
